list_models.py
import os
from google import genai
from dotenv import load_dotenv

def list_models():
    load_dotenv()
    api_key = os.getenv("GEMINI_API_KEY")
    client = genai.Client(api_key=api_key)
    
    print("Available Models:")
    print("-" * 50)
    try:
        for model in client.models.list():
            print(f"Name: {model.name} | Methods: {model.supported_methods}")
    except Exception as e:
        print(f"Error listing models: {e}")

if __name__ == "__main__":
    list_models()
