requirements.txt
fastapi
uvicorn
httpx
pydantic
pydantic-settings
python-dotenv
google-genai
aiogram>=3.0
